examples: don't link testchannels example with system libgstaudio
The testchannels audio test program is using -lgstaudio-0.10 to link
with libgstaudio which won't use the gstaudio library that was just
built but the one from the system. This is an issue since it means
we won't be testing the code from the current source tree, and it
also breaks the build when building on a system which don't have
a libgstaudio yet.

https://bugzilla.gnome.org/show_bug.cgi?id=652100