capsfilter: don't assume _get_caps still has to be
_set_caps only gets called when the buffer is actually pushed, so there
is a reasonably big window between when the initial caps are retrieved
and when the caps are set on our src pad. So we can't assume the not
having negotiated caps on our src pad means _get_caps still has to be
called.

Instead simply always suggest the new caps on buffer_alloc.