examples: Fixes for compiling with GTK+ 3.0
camerabing example had a missing ')' and the fix to build against GTK+3
that was introduced in commit 516c977c76ba4cd86a7147aab7416a33d76f870f
is also applied to the camerabin2 example.

The scaletempo demo was using API which has been removed, so that's
wrapped in a version check.